% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dicots_proportions}
\alias{dicots_proportions}
\title{Comparing biocontrol and herbicide for managing an invasive non-native plant species.}
\format{
An object of class \code{"tbl_df"} 

\describe{
\item{Treatment}{one of `B` (Biocontrol), `C` (Control), `H` (Herbicide), or `HB` (Herbicide and biocontrol)}
\item{Calluna_vulgaris}{proportion of cover in block}
\item{Exotic_dicots}{proportion of cover in block}
\item{Exotic_monocots}{proportion of cover in block}
\item{Native_dicots}{proportion of cover in block}
\item{Native_monocots}{proportion of cover in block}
\item{Bryophytes}{proportion of cover in block}
\item{Lichen}{proportion of cover in block}
\item{Clubmoss}{proportion of cover in block}
\item{bare}{proportion of cover in block}
\item{litter}{proportion of cover in block}
}
}
\source{
\href{https://besjournals.onlinelibrary.wiley.com/doi/full/10.1111/1365-2664.13691}{Journal of Applied Ecology}
}
\usage{
data(dicots_proportions)
}
\description{
Data from a study conducted to compare two stakeholder‐selected control methods for heather,
a European plant invading native ecosystems in and adjoining Tongariro National Park in New Zealand
}
\examples{
data(dicots_proportions)
\donttest{dicots_proportions}
}
\keyword{datasets}
