% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{howls}
\alias{howls}
\title{Recorded howls of the Indian wolf}
\format{
An object of class \code{"tbl_df"} with 132 rows and 14 columns
}
\source{
\href{https://github.com/bhlabwii/wolf_howlID}{GitHub}
}
\usage{
data(howls)
}
\description{
Data from a survey conducted recording the howls of Indian wolves.
}
\examples{
data(howls)
\donttest{howls}
}
\references{
Sadhukhan, S., Root-Gutteridge, H. & Habib, B. Identifying unknown Indian wolves by their
distinctive howls: its potential as a non-invasive survey method. Sci Rep 11, 7309 (2021)
(\href{https://www.nature.com/articles/s41598-021-86718-w}{Scientific Reports})
}
\keyword{datasets}
