% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lobster}
\alias{lobster}
\title{Influence of predator identity on the strength of predator avoidance responses in lobsters}
\format{
An object of class \code{"tbl_df"} with 159 rows and 2 columns

\describe{
  \item{size}{the length of the lobster’s carapace, to the nearest 3 mm.}
  \item{survived}{this takes the value 1 if the lobster survived, and the value 0 if it did not.}
}
}
\source{
\href{https://www.sciencedirect.com/science/article/pii/S0022098115000039}{Journal of Experimental Marine Biology and Ecology}
}
\usage{
data(lobster)
}
\description{
Data from a study conducted to investigate how a juvenile lobster’s size was related to its
vulnerability to predation.
}
\examples{
data(lobster)
\donttest{lobster}
}
\references{
Wilkinson, E. B., Grabowski, J. H., Sherwood, G. D., and Yund, O. Y.
Influence of predator identity on the strength of predator avoidance responses in lobsters (2015)
(\href{https://www.sciencedirect.com/science/article/pii/S0022098115000039}{Journal of Experimental Marine Biology and Ecology})
}
\keyword{datasets}
