% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lizards}
\alias{lizards}
\title{Behavioral response of the Mediterranean lizard to smoke}
\format{
An object of class \code{"tbl_df"} with 216 rows and 13 columns

\describe{
  \item{Code}{individual lizard ID}
  \item{order_treatment}{order the treatments were applied}
  \item{treatment}{either `Control` or `Smoke`}
  \item{post_activity_s}{activity in seconds one minute after the treatment application }
}
}
\source{
\href{https://doi.org/10.1093/beheco/arab010}{Behavioral Ecology}
}
\usage{
data(lizards)
}
\description{
Data from a study conducted to investigate whether the Mediterranean lizard (Psammodromus algirus)
could recognise the threat of fire by detecting smoke and triggering a behavioral
response that enhances survival in fire-prone ecosystems.
}
\examples{
data(lizards)
\donttest{lizards}
}
\references{
Lola Álvarez-Ruiz, Josabel Belliure, Juli G Pausas, Fire-driven behavioral
response to smoke in a Mediterranean lizard, Behavioral Ecology, 2021
(\href{doi:10.5061/dryad.xpnvx0kdv}{Behavioral Ecology})
}
\keyword{datasets}
